ecrivain fonctionnel

+++ src/main/java/org/example/projetjardinage/model/journal/mesures/MesureNumerique.java
@@ -18,7 +18,7 @@ public class MesureNumerique extends Mesure {
     public Float getValue() { return value; }
     public void setValue(float value) { this.value = value; }
 
-    public void setValueLect(String val) {this.value = Integer.parseInt(val);}
+    public void setValueLect(String val) {this.value = Float.parseFloat(val);}
 
     public String getUnit() { return unit; }
     public void setUnit(String unit) { this.unit = unit; }
